+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## 12.0.0
+
+Makes it so the scripts are loaded using `SCRIPT LOAD` so they aren't sent
+to redis each time.
+
+Adds disabled-by-default feature that enables milisecond time precission.
+
 ## 10.0.0 - 2024-05-08
 
-Change format from queue names for compatibiliti with the original Nodejs version of the crate. 
+Change format from queue names for compatibiliti with the original Nodejs version of the crate.
 Change details in here: https://github.com/DavidBM/rsmq-async-rs/pull/20
 
 ### Changed